requirements.txt
python-telegram-bot==20.7
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
sqlite3